requirements.txt
python-telegram-bot==13.12
environs==9.5.0
google-cloud-dialogflow==2.14.0
vk-api==11.9.8